split the debug prompt/response in a debug mixin
